Mudando os arquivos da pasta v2 para a principal
