+++ quadribol-api/src/main/java/quadribolapi/domain/Acomodacao.java
