+++ quadribol-api/src/main/java/quadribolapi/domain/Administrador.java
