+++ quadribol-api/src/main/java/quadribolapi/domain/Arbitro.java
@@ -11,46 +11,47 @@ public class Arbitro {
 	
 	@Id
 	@GeneratedValue(strategy = GenerationType.IDENTITY)
-	private final Long id;
+	private final Long ID_ARBITRO;
 	
 	@Column(nullable = false)
-	private String nome;
+	private String nomeArbitro;
 
     @Column(nullable = false)
-	private int idade;
+	private int idadeArbitro;
 
-	public Arbitro(Long ide, String nom, int idad) {
-        this.id = ide;
-        this.nome = nom;
-        this.idade = idad;
+	private static int contadorArbitro = 0;
+
+	public Arbitro(String nome, int idade){
+        this.idArbitro = contadorArbitro;
+		contadorArbitro++;
+        this.nomeArbitro = nome;
+        this.idadeArbitro = idade;
     }
     
-    public Long getId() {
-		return id;
+    public Long getIdArbitro() {
+		return this.ID_ARBITRO;
 	}
 
-    public String getNome() {
-		return this.nome;
+	public String getNomeArbitro() {
+		return this.nomeArbitro;
 	}
 
-	public void setNome(String novon) {
-		this.nome = novon;
+	public void setNomeArbitro(String novoNome) {
+		this.nomeArbitro = novoNome;
 	}
 
-    public int getIdade() {
-		return this.idade;
-	}
-
-    public void setIdade(int novai) {
-		this.idade = novai;
-	}
-
-    public void exibirInfoArbitro() {
-        
-        System.out.printf("ID: 0x%016X", this.getId());
-        System.out.printf("Nome: %s", this.getNome());
-        System.out.printf("Idade: %d", this.getIdade());
-        
+    public int getIdadeArbitro() {
+        return this.idadeArbitro;
     }
+
+    public void setIdadeArbitro(int novaIdade) {
+        this.idadeArbitro = novaIdade;
+    }
+
+	public void exibirInfoArbitro() {
+		System.out.printf("ID: 0x%016X", this.getIdArbitro());
+        System.out.printf("Nome: %s", this.getNomeArbitro());
+        System.out.printf("Idade: %d", this.getIdadeArbitro());
+	}
 
 }
