+++ quadribol-api/src/main/java/quadribolapi/domain/Arena.java
