+++ quadribol-api/src/main/java/quadribolapi/domain/GerenciadorUsuarios.java
