+++ quadribol-api/src/main/java/quadribolapi/domain/Jogador.java
@@ -1,47 +1,83 @@
 package quadribolapi.domain;
 
+import javax.persistence.Column;
 import javax.persistence.Entity;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
 import javax.persistence.Id;
 
 @Entity
 public class Jogador {
 	
 	@Id
-	private int id;
-	private String nome;
-	private int idade;
+	@GeneratedValue(strategy = GenerationType.IDENTITY)
+	private final Long ID_JOGADOR;
+	
+	@Column(nullable = false)
+	private String nomeJogador;
+
+    @Column(nullable = false)
+	private int idadeJogador;
+
+    @Column(nullable = false)
 	private String posicao;
-	
-	public Jogador(int id, String nome, int idade, String posicao) {
-		super();
-		this.id = id;
-		this.nome = nome;
-		this.idade = idade;
-		this.posicao = posicao;
+
+    @Column(nullable = false)
+	private int pontos;
+
+    private static int contadorJogador = 0;
+
+	public Jogador(String nome, int idade, String posicao, int pontos){
+        this.idJogador = contadorJogador;
+        contadorJogador++;
+        this.nomeJogador = nome;
+        this.idadeJogador = idade;
+        this.posicao = posicao;
+        this.pontos = pontos;
+    }
+    
+    public Long getIdJogador() {
+		return this.ID_JOGADOR;
 	}
-	
-	public int getId() {
-		return id;
+
+	public String getNomeJogador() {
+		return this.nomeJogador;
 	}
-	
-	public String getNome() {
-		return nome;
+
+	public void setNomeJogador(String novoNome) {
+		this.nomeJogador = novoNome;
 	}
-	public void setNome(String nome) {
-		this.nome = nome;
-	}
-	public int getIdade() {
-		return idade;
-	}
-	public void setIdade(int idade) {
-		this.idade = idade;
-	}
-	public String getPosicao() {
-		return posicao;
-	}
-	public void setPosicao(String posicao) {
-		this.posicao = posicao;
-	}
-	
-	
+
+    public int getIdadeJogador() {
+        return this.idadeJogador;
+    }
+
+    public void setIdadeJogador(int novaIdade) {
+        this.idadeJogador = novaIdade;
+    }
+
+    public String getPosicao() {
+        return this.posicao;
+    }
+
+    public void setPosicao(String novaPosicao) {
+        this.posicao = novaPosicao;
+    }
+
+    public int getPontos() {
+        return this.pontos;
+    }
+
+    public void setPontos(int novoPontos) {
+        this.pontos = novoPontos;
+    }
+
+    public void exibirInfoJogador() {
+        System.out.printf("ID: 0x%016X", this.getIdJogador());
+        System.out.printf("Nome: %s", this.getNomeJogador());
+        System.out.printf("Idade: %d", this.getIdadeJogador());
+        System.out.printf("Posicao: %s", this.getPosicao());
+        System.out.printf("Pontos: %d", this.getPontos());
+    }
+
 }
